+++ resolwe_bio/tests/workflows/test_geo_import.py
@@ -3,119 +3,12 @@ from pathlib import Path
 from django.test import LiveServerTestCase
 
 from resolwe.flow.models import Collection, Data
-from resolwe.flow.models.annotations import (
-    AnnotationField,
-    AnnotationGroup,
-    AnnotationType,
-)
 from resolwe.test import tag_process, with_resolwe_host
 
 from resolwe_bio.utils.test import BioProcessTestCase
 
 
 class GeoImportTestCase(BioProcessTestCase, LiveServerTestCase):
-    def setUp(self):
-        """Initialize annotation groups and fields."""
-        super().setUp()
-
-        general_group = AnnotationGroup.objects.get(name="general")
-
-        AnnotationField.objects.create(
-            name="annotator",
-            sort_order=1,
-            group=general_group,
-            type=AnnotationType.STRING.value,
-        )
-
-        AnnotationField.objects.create(
-            name="description",
-            sort_order=1,
-            group=general_group,
-            type=AnnotationType.STRING.value,
-        )
-
-        biospecimen_group = AnnotationGroup.objects.create(
-            name="biospecimen_information", sort_order=1
-        )
-
-        AnnotationField.objects.create(
-            name="experimental_model",
-            sort_order=1,
-            group=biospecimen_group,
-            type=AnnotationType.STRING.value,
-        )
-
-        AnnotationField.objects.create(
-            name="source",
-            sort_order=1,
-            group=biospecimen_group,
-            type=AnnotationType.STRING.value,
-        )
-
-        cell_line_group = AnnotationGroup.objects.create(
-            name="cell_line_information", sort_order=1
-        )
-
-        AnnotationField.objects.create(
-            name="cell_line_name",
-            sort_order=1,
-            group=cell_line_group,
-            type=AnnotationType.STRING.value,
-        )
-
-        AnnotationField.objects.create(
-            name="cell_type",
-            sort_order=1,
-            group=cell_line_group,
-            type=AnnotationType.STRING.value,
-        )
-
-        AnnotationField.objects.create(
-            name="treatment_protocol",
-            sort_order=1,
-            group=cell_line_group,
-            type=AnnotationType.STRING.value,
-        )
-
-        sample_details_group = AnnotationGroup.objects.create(
-            name="sample_details", sort_order=1
-        )
-
-        AnnotationField.objects.create(
-            name="assay_type",
-            sort_order=1,
-            group=sample_details_group,
-            type=AnnotationType.STRING.value,
-        )
-
-        AnnotationField.objects.create(
-            name="extract_protocol",
-            sort_order=1,
-            group=sample_details_group,
-            type=AnnotationType.STRING.value,
-        )
-
-        AnnotationField.objects.create(
-            name="growth_protocol",
-            sort_order=1,
-            group=sample_details_group,
-            type=AnnotationType.STRING.value,
-        )
-
-        AnnotationField.objects.create(
-            name="library_type",
-            sort_order=1,
-            group=sample_details_group,
-            type=AnnotationType.STRING.value,
-        )
-
-        AnnotationField.objects.create(
-            name="platform",
-            sort_order=1,
-            group=sample_details_group,
-            type=AnnotationType.STRING.value,
-        )
-
     @with_resolwe_host
     @tag_process("geo-import")
     def test_dss_geo(self):
@@ -161,42 +54,6 @@ class GeoImportTestCase(BioProcessTestCase, LiveServerTestCase):
             ],
         )
         sra = Data.objects.filter(process__slug="import-sra-single").last()
-        sample = sra.entity
-
-        self.assertEqual(sample.annotations.count(), 12)
-
-        self.assertAnnotation(sample, "general.species", "Homo sapiens")
-
-        self.assertAnnotation(sample, "general.annotator", "Lin He")
-        self.assertAnnotation(sample, "general.description", "ING5 knockdown")
-
-        self.assertAnnotation(
-            sample, "biospecimen_information.experimental_model", "cell_line"
-        )
-        self.assertAnnotation(sample, "biospecimen_information.source", "HepG2 cells")
-
-        self.assertAnnotation(sample, "cell_line_information.cell_line_name", "HepG2")
-        self.assertAnnotation(
-            sample,
-            "cell_line_information.treatment_protocol",
-            (
-                "HepG2 cells were transfected with vector or FLAG-JFK or treated with control "
-                "siRNA or ING5 siRNA."
-            ),
-        )
-
-        self.assertAnnotation(sample, "sample_details.assay_type", "rna-seq")
-        self.assertAnnotation(
-            sample,
-            "sample_details.extract_protocol",
-            (
-                "Total mRNAs were isolated with Trizol reagents (Invitrogen) for cDNA synthesis, "
-                "library construction, and sequencing using HiSeq 2500. RNA libraries were "
-                "prepared for sequencing using standard Illumina protocols."
-            ),
-        )
-        self.assertAnnotation(sample, "sample_details.library_type", "total_rna")
-        self.assertAnnotation(sample, "sample_details.platform", "hiseq_2000")
 
         # Non-existant GSE series.
         wrong = self.run_process(
@@ -323,51 +180,6 @@ class GeoImportTestCase(BioProcessTestCase, LiveServerTestCase):
             ],
         )
 
-        sample = sra.entity
-
-        self.assertEqual(sample.annotations.count(), 11)
-
-        # general
-        self.assertAnnotation(sample, "general.species", "Homo sapiens")
-        self.assertAnnotation(sample, "general.annotator", "Matthew Weirauch")
-        self.assertAnnotation(
-            sample, "general.description", "ChIP_EBNA2_GM12878_rep2-E00457"
-        )
-
-        # biospecimen_information
-        self.assertAnnotation(
-            sample, "biospecimen_information.experimental_model", "cell_line"
-        )
-        self.assertAnnotation(
-            sample, "biospecimen_information.source", "GM12878 (B-Lymphocyte) LCL"
-        )
-
-        # cell_line_information
-        self.assertAnnotation(
-            sample, "cell_line_information.cell_line_name", "GM12878 (B-Lymphocyte) LCL"
-        )
-        self.assertAnnotation(sample, "cell_line_information.treatment_protocol", "")
-
-        # sample_details
-        self.assertAnnotation(sample, "sample_details.assay_type", "chip-seq")
-        self.assertAnnotation(
-            sample,
-            "sample_details.extract_protocol",
-            (
-                "Cells were crosslinked and nuclei were sonicated as described previously "
-                "(Lu et al. 2015). Libraries were prepared via ChIPmentation (Schmidl et al. "
-                "2015)."
-            ),
-        )
-        self.assertAnnotation(
-            sample,
-            "sample_details.growth_protocol",
-            (
-                "Cells were cultured in 10% FBS supplemented RPMI 1640 medium for 2 weeks."
-            ),
-        )
-        self.assertAnnotation(sample, "sample_details.library_type", "genomic_dna")
-
     @with_resolwe_host
     @tag_process("geo-import")
     def test_geo_ena(self):
@@ -414,57 +226,3 @@ class GeoImportTestCase(BioProcessTestCase, LiveServerTestCase):
                 },
             ],
         )
-
-        sample = sra.entity
-
-        self.assertEqual(sample.annotations.count(), 7)
-
-        # general
-        self.assertAnnotation(sample, "general.annotator", "ArrayExpress EBI")
-        self.assertAnnotation(
-            sample, "general.description", "Provider: EMBL_Heidelberg"
-        )
-
-        # biospecimen_information
-        self.assertAnnotation(
-            sample,
-            "biospecimen_information.source",
-            "E_coli_K12_MG1655_RNAseq_LB_Transition_to_stationary",
-        )
-
-        # sample_details
-        self.assertAnnotation(sample, "sample_details.assay_type", "rna-seq")
-
-        self.assertAnnotation(
-            sample,
-            "sample_details.extract_protocol",
-            (
-                "nucleic_acid_extraction | To prepare cells for RNA "
-                "extraction, 100 ml of fresh LB was inoculated 1:200 from an overnight culture "
-                "in a 250 ml flask and incubated with shaking at 180 r.p.m. in a New Brunswick "
-                "C76 waterbath at 37C. Two biological replicates were performed for each strain "
-                "and samples were taken at early-exponential, mid-exponential, "
-                "transition-to-stationary and stationary phase. The cells were pelleted by "
-                "centrifugation (10000 g, 10 min, 4¡C), washed in 1xPBS and pellets were "
-                "snap-frozen and stored at -80C until required. RNA was extracted using Trizol "
-                "Reagent (Invitrogen) according to the manufacturer's protocol until the "
-                "chloroform extraction step. The aqueous phase was then loaded onto mirVanaTM "
-                "miRNA Isolation kit (Ambion Inc.) columns and washed according to the "
-                "manufacturer's protocol. Total RNA was eluted in 50µl of RNAase free water. "
-                "The concentration was then determined using a Nanodrop ND-1000 machine (NanoDrop "
-                "Technologies), and RNA quality was tested by visualization on agarose gels and "
-                "by Agilent 2100 Bioanalyser (Agilent Technologies). sequencing | Standard "
-                "Illumina protocol for cDNA sequencing"
-            ),
-        )
-        self.assertAnnotation(
-            sample,
-            "sample_details.growth_protocol",
-            (
-                "grow | The E. coli K-12 MG1655 bacterial strains used in this "
-                "work are the following: E. coli MG1655 (F- lambda- ilvG- rfb-50 rph-1). "
-                "Luria-Bertani (0.5% NaCl) broth and agar (15 g/liter) were used for routine "
-                "growth."
-            ),
-        )
-        self.assertAnnotation(sample, "sample_details.library_type", "total_rna")