+++ docs/CHANGELOG.rst
@@ -22,6 +22,7 @@ Changed
 - Rewrite ``expression-aggregator`` process to Python and make it
   compatible with the new annotation model
 - Use GRCh38.109 genome annotation version in snpEff processes
+- Remove sample annotation functionality from ``geo-import`` process
 
 Fixed
 -----